studentCommands/NewStudentCommand: Fix incorrect input splitting and validation
The NewStudentCommand was incorrectly splitting the user input when processing student names, causing issues with names that contain multiple parts (e.g., "John Doe"). Additionally, it was not properly checking the number of parts in the command, leading to potential errors when fields were missing.

To address these issues, I:
- Fixed the logic to properly handle multi-part student names splitting by commas instead of spaces.
- Added validation to ensure the correct number of input parts before processing the student data.

This fix ensures that the command now correctly parses and handles multi-part student names and validates input length before proceeding.
